Separated Shortcode classes into own files
