+++ gulpfile.ts/build.ts
@@ -7,6 +7,7 @@ const gulp = require( 'gulp' );
 const cleanCSS = require( 'gulp-clean-css' );
 const composer = require( 'gulp-uglify/composer' );
 const concat = require( 'gulp-concat' );
+const inject = require( 'gulp-inject-string' );
 const merge = require( 'merge-stream' );
 const rename = require( 'gulp-rename' );
 const replace = require( 'gulp-replace' );
@@ -208,13 +209,17 @@ gulp.task( 'build:ts:admin', function() {
 } );
 
 gulp.task( 'build:ts:frontend', function() {
-	function bundle( name: string, sources: Array<string> ): NodeJS.ReadWriteStream {
+	function bundle( name: string, sources: Array<string>, jQuery = false ): NodeJS.ReadWriteStream {
 		const tsProject = ts.createProject( 'tsconfig.json' );
-		return gulp.src( sources.concat( [ 'src/d.ts/*.d.ts' ] ) )
+		let ret = gulp.src( sources.concat( [ 'src/d.ts/*.d.ts' ] ) )
 			.pipe( tsProject() )
 			.js
-			.pipe( concat( name + '.min.js' ) )
-			.pipe( minify( { ie8: true } ) )
+			.pipe( concat( name + '.min.js' ) );
+		if ( jQuery ) {
+			ret = ret.pipe( inject.prepend( 'jQuery( document ).ready( function( $ ) {\n' ) )
+				.pipe( inject.append( '} );\n' ) );
+		}
+		return ret.pipe( minify( { ie8: true } ) )
 			.pipe( gulp.dest( 'dist/frontend/js/' ) );
 	}
 
@@ -233,9 +238,11 @@ gulp.task( 'build:ts:frontend', function() {
 		] ),
 		bundle( 'shortcode', [
 			'src/ts/isError.ts',
-			'src/ts/frontend/shortcode/queryParameter.ts',
 			'src/ts/frontend/shortcode.ts',
-		] )
+			'src/ts/frontend/shortcode/queryParameter.ts',
+			'src/ts/frontend/shortcode/Shortcode.ts',
+			'src/ts/frontend/shortcode/ShortcodeRegistry.ts',
+		], true )
 	);
 } );
